+++ packages/abstract-image/src/exporters/__tests__/react-svg-export-image/test-defs/react-svg-dashed-line.tsx
@@ -6,7 +6,9 @@ const components = [
     AbstractImage.createPoint(10, 100),
     AbstractImage.createPoint(40, 70),
     AbstractImage.black,
-    2
+    2,
+    undefined,
+    AbstractImage.createDashStyle([10, 5], 3)
   ),
 ];
 const image = AbstractImage.createAbstractImage(
@@ -19,8 +21,8 @@ const image = AbstractImage.createAbstractImage(
 const svg = AbstractImage.createReactSvg(image);
 
 export const test: ExportTestDef = {
-  name: "react svg line",
+  name: "react svg dashed line",
   abstractImage: svg,
   expectedSerializedJsx:
-    '{"type":"svg","key":null,"ref":null,"props":{"id":"ai_root","width":"400px","height":"400px","viewBox":"0 0 400 400","children":[{"key":"0","ref":null,"props":{"component":{"type":"line","start":{"x":10,"y":100},"end":{"x":40,"y":70},"strokeColor":{"a":255,"r":0,"g":0,"b":0},"strokeThickness":2,"strokeDashStyle":{"dashes":[],"offset":0}}},"_owner":null,"_store":{}}]},"_owner":null,"_store":{}}',
+    '{"type":"svg","key":null,"ref":null,"props":{"id":"ai_root","width":"400px","height":"400px","viewBox":"0 0 400 400","children":[{"key":"0","ref":null,"props":{"component":{"type":"line","start":{"x":10,"y":100},"end":{"x":40,"y":70},"strokeColor":{"a":255,"r":0,"g":0,"b":0},"strokeThickness":2,"strokeDashStyle":{"dashes":[10,5],"offset":3}}},"_owner":null,"_store":{}}]},"_owner":null,"_store":{}}',
 };