+++ packages/abstract-image/src/model/dash-style.ts
@@ -1,0 +1,13 @@
+export interface DashStyle {
+  readonly dashes: ReadonlyArray<number>;
+  readonly offset: number;
+}
+
+export function createDashStyle(dashes: ReadonlyArray<number>, offset: number = 0): DashStyle {
+  return {
+    dashes: dashes,
+    offset: offset,
+  };
+}
+
+export const solidLine = createDashStyle([]);