+++ packages/abstract-image/src/model/index.ts
@@ -3,3 +3,4 @@ export * from "./color";
 export * from "./component";
 export * from "./point";
 export * from "./size";
+export * from "./dash-style";